clean list
