+++ uikit/domql/List/GroupListWithSearch.js
@@ -1,0 +1,4 @@
+'use strict'
+
+export const GroupListWithSearch = {
+}
